// src/types/defaults-types.ts
export interface A11yToolbarDefaultOptions {
  showReadSpeakerButton?: boolean;
  showTextSizeButton?: boolean;
  showContrastButton?: boolean;
  showPrintButton?: boolean;
  showLanguageButton?: boolean;
  readSpeakerID?: string;
  iconOptions?: {
    readSpeakerIcon?: string;
    textSizeIcon?: string;
    contrastIcon?: string;
    printIcon?: string;
    languageIcon?: string;
  };
}